% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gremlin.R
\name{fixef.gremlin}
\alias{fixef.gremlin}
\alias{fixef}
\alias{fixed.effects}
\title{Fixed Effect Estimates of \code{class} \sQuote{gremlin}}
\usage{
\method{fixef}{gremlin}(object, add.dropped = FALSE, ...)
}
\arguments{
\item{object}{An object of \code{class} \sQuote{gremlin}.}

\item{add.dropped}{A \code{logical} value indicating whether fixed effects dropped by
gremlin, due to rank deficiencies in the fixed effect design matrix, should
be included with \code{NA} values.}

\item{\dots}{Additional arguments.}
}
\value{
A \code{numeric} vector of fixed effect estimates.
}
\description{
Extracts the fixed effect estimates from a model of \code{class} \sQuote{gremlin}.
}
\examples{
fixef(gremlin(WWG11 ~ sex - 1, random = ~ sire, data = Mrode11))
}
\author{
\email{matthewwolak@gmail.com}
}
