% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlin.R, R/mkModMats.R
\name{gremlin}
\alias{gremlin}
\alias{gremlinR}
\alias{gremlinSetup}
\alias{mkModMats}
\alias{update}
\alias{getCall.gremlin}
\alias{update.gremlin}
\alias{is.gremlin}
\alias{is.grMod}
\title{Mixed-effect modeling functions.}
\usage{
gremlin(formula, random = NULL, rcov = ~units, data = NULL,
  ginverse = NULL, Gstart = NULL, Rstart = NULL, Bp = NULL,
  maxit = 20, algit = NULL, vit = 10, v = 1, cctol = c(5e-04,
  1e-08, 0.001, NULL), ezero = 1e-08, ...)

gremlinR(formula, random = NULL, rcov = ~units, data = NULL,
  ginverse = NULL, Gstart = NULL, Rstart = NULL, Bp = NULL,
  maxit = 20, algit = NULL, vit = 10, v = 1, cctol = c(5e-04,
  1e-08, 0.001, NULL), ezero = 1e-08, ...)

\method{getCall}{gremlin}(x, ...)

update.gremlin(object, ...)

gremlinSetup(formula, random = NULL, rcov = ~units, data = NULL,
  ginverse = NULL, Gstart = NULL, Rstart = NULL, Bp = NULL,
  maxit = 20, algit = NULL, vit = 10, v = 1, cctol = c(5e-04,
  1e-08, 0.001, NULL), ezero = 1e-08, ...)

\method{is}{gremlin}(x)

\method{is}{grMod}(x)

mkModMats(formula, random = NULL, rcov = ~units, data = NULL,
  subset = NULL, ginverse = NULL, na.action = na.pass,
  offset = NULL, contrasts = NULL, Xsparse = TRUE, ...)
}
\arguments{
\item{formula}{A \code{formula} for the response variable(s) and fixed effects.}

\item{random}{A \code{formula} for the random effects.}

\item{rcov}{A \code{formula} for the residual covariance structure.}

\item{data}{A \code{data.frame} in which to look for the terms in
\code{formula}, \code{random}, and \code{rcov}.}

\item{ginverse}{A \code{list} of (preferably sparse) inverse matrices that
are proportional to the covariance structure of the random effects.
The name of each element in the list should match a column in \code{data}
that is associated with a random term. All levels of the random term should
appear as \code{rownames} for the matrices.}

\item{Gstart}{A \code{list} of starting (co)variance values for the the
G-structure or random terms.}

\item{Rstart}{A \code{list} of starting (co)variance values for the
R-structure or residual terms.}

\item{Bp}{A prior specification for fixed effects.}

\item{maxit}{An \code{integer} specifying the maximum number of likelihood
iterations.}

\item{algit}{A \code{character} vector of length 1 or more or an expression
to be evaluated that specifies the algorithm to use for proposing
(co)variances in the next likelihood iteration.}

\item{vit}{An \code{integer} value specifying the verbosity of screen output
on each iteration. A value of zero gives no iteration specific output and
larger values increase the amount of information printed on the screen.}

\item{v}{An \code{integer} value specifying the verbosity of screen output
regarding the model fitting process. A value of zero gives no details and
larger values increase the amount of information printed on the screen.}

\item{cctol}{Convergence criteria tolerances. TODO}

\item{ezero}{Effective zero to be used, values less than this number are
treated as zero and fixed to this value.}

\item{\dots}{Additional arguments to be passed to control the model fitting.}

\item{subset}{An expression for the subset of \code{data} to use.}

\item{na.action}{What to do with NAs.}

\item{offset}{Should an offset be specified.}

\item{contrasts}{Specify the type of contrasts for the fixed effects.}

\item{Xsparse}{Should sparse matrices be used for the fixed effects design
matrix.}
}
\description{
Fit and setup functions for linear mixed-effect model (Gaussian responses).
}
\section{Functions}{
\itemize{
\item \code{mkModMats}: Generates model matrices.
}}

\examples{
 TODO XXX XXX Make example of each modular component working with next
  mod11 <- gremlinRmod(WWG11 ~ sex - 1,
  	random = ~ calf,
  	data = Mrode11,
  	Gstart = matrix(0.2), Rstart = matrix(0.4),
  	maxit = 10, v = 2)

}
\references{
Meyer.
Henderson.
Mrode. 2005.
}
\author{
\email{matthewwolak@gmail.com}
}
