% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlinControl.R
\name{gremlinControl}
\alias{gremlinControl}
\title{Advanced Options for Mixed-effect modeling functions.}
\usage{
gremlinControl(cctol = c(5e-04, 1e-08, 0.001, NULL), ezero = 1e-08,
  lambda = TRUE, algorithm = NULL, algArgs = list())
}
\arguments{
\item{cctol}{Convergence criteria tolerances. TODO}

\item{ezero}{Effective zero to be used, values less than this number are
treated as zero and fixed to this value.}

\item{lambda}{A \code{logical} indicating whether a residual variance should
be factored out of the mixed model equations.}

\item{algorithm}{A \code{character} naming the function to use to decide
subsequent parameters in the REML iterations.}

\item{algArgs}{A \code{list} of function arguments to be given to functions
named in the \code{algorithm} argument.}
}
\value{
A \code{list} of class \code{gremlinControl} to be used by
  \code{gremlinSetup} and later functions when fitting the model.
}
\description{
Change default settings for gremlin models.
}
\examples{
  str(gremlinControl())

}
\references{
Meyer. Convergence Checks
}
\author{
\email{matthewwolak@gmail.com}
}
