% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gremlin.R
\name{logLik.gremlin}
\alias{logLik.gremlin}
\title{Methods to extract log-likelihood and information criterion of a gremlin
model.}
\usage{
\method{logLik}{gremlin}(object, ...)
}
\arguments{
\item{object}{An object of \code{class} \sQuote{gremlin}.}

\item{\dots}{Additional arguments to be passed to control the model fitting.}
}
\value{
A \code{numeric} value for the log-likelihood.
}
\description{
Extracts the log-likelihood from a gremlin model fit.
}
\examples{
mod11 <- gremlinR(WWG11 ~ sex - 1,
random = ~ calf,
data = Mrode11,
Gstart = matrix(0.1), Rstart = matrix(0.4),
maxit = 50, v = 2, algit = "EM")
logLik(mod11)
}
\author{
\email{matthewwolak@gmail.com}
}
