\name{Mrode11}
\alias{Mrode11}
\docType{data}
\title{
Weight Gain data
}
\description{
Data from chapter 11 in Mrode 2005
}
\usage{data("Mrode11")}
\format{
  A data frame with 5 observations on the following 5 variables.
  \describe{
    \item{\code{calf}}{a factor with levels \code{4} \code{5} \code{6} \code{7} \code{8}}
    \item{\code{dam}}{a factor with levels \code{2} \code{5} \code{6}}
    \item{\code{sire}}{a factor with levels \code{1} \code{3} \code{4}}
    \item{\code{sex}}{a factor with levels \code{male} \code{female}}
    \item{\code{WWG11}}{a numeric vector}
  }
}

\source{
Mrode. 2005.
}

\examples{
data(Mrode11)
}
\keyword{datasets}
