% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlin.R
\name{remlIt}
\alias{remlIt}
\alias{remlIt.default}
\title{Mixed-effect model Restricted Maximul Likelihood (REML) iterations.}
\usage{
remlIt(grMod, ...)

\method{remlIt}{default}(grMod, ...)
}
\arguments{
\item{grMod}{A gremlin model of class \code{grMod}. See \code{\link{gremlin}}
or \code{\link{gremlinSetup}} for the functions constructing an object
of class \code{grMod}.}

\item{\dots}{Additional arguments to be passed to control the model fitting.}
}
\value{
A \code{list} containing an object of class \code{grMod} and
  \code{matrix} containing details of the REML iterations (object
  \code{itMat}). See \code{\link{gremlin}} for descriptions of \code{grMod}
  and \code{itMat} objects.
}
\description{
Conduct REML iterations to estimate (co)variance parameters of a linear
  mixed-effect model (Gaussian responses).
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method
}}

\examples{
 TODO XXX XXX Make example of each modular component working with next
  mod11 <- gremlinSetup(WWG11 ~ sex - 1,
  	random = ~ calf,
  	data = Mrode11,
  	Gstart = matrix(0.2), Rstart = matrix(0.4),
  	maxit = 10, v = 2)
  REMLout <- remlIt(mod11)

}
\author{
\email{matthewwolak@gmail.com}
}
