% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gremlin.R
\name{residuals.gremlin}
\alias{residuals.gremlin}
\alias{resid.gremlin}
\title{Residuals of \code{class} \sQuote{gremlin}}
\usage{
\method{residuals}{gremlin}(object, type = "response", scaled = FALSE, ...)
}
\arguments{
\item{object}{An object of \code{class} \sQuote{gremlin}.}

\item{type}{The type of residuals which should be returned. Only implement
\dQuote{response} currently. Can be abbreviated.}

\item{scaled}{Logical value indicating whether to scale residuals by the
residual standard deviation.}

\item{\dots}{Additional arguments.}
}
\value{
A \code{numeric} vector of residuals.
}
\description{
Residuals of \code{class} \sQuote{gremlin}.
}
\examples{
mod11 <- gremlinR(WWG11 ~ sex - 1,
random = ~ calf,
data = Mrode11,
Gstart = matrix(0.1), Rstart = matrix(0.4),
maxit = 50, v = 2, algit = "EM")
residuals(mod11)
}
\author{
\email{matthewwolak@gmail.com}
}
