% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gremlin.R
\name{runtime}
\alias{runtime}
\title{Time to execute the gremlin model}
\usage{
runtime(object, ...)
}
\arguments{
\item{object}{An object of \code{class} \sQuote{gremlin}.}

\item{\dots}{Further arguments to be passed to the methods.}
}
\value{
A numeric of class \code{\sQuote{difftime}} with an attribute of units
(e.g., seconds or minutes).
}
\description{
Extract the length of time to fit the model.
}
\examples{
mod11 <- gremlinR(WWG11 ~ sex - 1,
random = ~ calf,
data = Mrode11, algit = "EM")
gremlin:::runtime(mod11)
}
\author{
\email{matthewwolak@gmail.com}
}
