\name{stTrans}
\alias{stTrans}
\title{
Transformation of starting parameters
}
\description{
Transform start parameters into format \code{gremlin} expects}
\usage{
stTrans(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x }{A \code{list} of starting parameters.}
}

\value{
A sparse \sQuote{dsCMatrix}
}

\author{\email{matthewwolak@gmail.com}}

