% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algChk.R
\name{algChk}
\alias{algChk}
\title{Optimization Algorithm Checks.}
\usage{
algChk(algit, maxit, ctrl, mc)
}
\arguments{
\item{algit}{A character \code{vector} of algorithms for each iteration.}

\item{maxit}{An \code{integer} indicating the number of REML iterations.}

\item{ctrl}{A \code{list} of arguments set by \code{gremlinControl}.}

\item{mc}{A model \code{call}.}
}
\value{
A \code{list} containing a vector \code{algit} specifying the currently
  implemented optimization algorithms for each iteration along with a vector
  each containing the type of first (\code{fdit}) and second (\code{sdit})
  derivatives for each iteration (else \code{NA} if either is not applicable).
}
\description{
Check and/or set optimization algorithms to use. Intended for internal use
within gremlin
}
\author{
\email{matthewwolak@gmail.com}
}
