% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlin.R
\docType{package}
\name{gremlin-package}
\alias{gremlin-package}
\alias{_PACKAGE}
\title{Mixed-Effects REML incoporating Generalized Inverses}
\description{
Fit linear mixed-effects models using restricted (or residual) maximum
likelihood (REML) and with generalized inverse matrices to specify covariance
structures for random effects. In particular, the package is suited to fit
quantitative genetic mixed models, often referred to as 'animal models'.
Implements the average information algorithm as the main tool to maximize the
restricted likelihood, but with other algorithms available.
}
\details{
The package also implements the average information algorithm to efficiently
maximize the log-likelihood (Thompson & Johnson 1995; Gilmour et al. 1995;
Meyer & Smith 1996). The average information algorithm combined with sparse
matrix techniques can potentially make model fitting very efficient.
}
\examples{
#TODO: simple examples of the most important functions
\dontrun{
  library(nadiv)
  Ainv <- makeAinv(Mrode3[-c(1:2), 1:3])$Ainv
  mod11 <- gremlin(WWG11 ~ sex - 1,
random = ~ calf,
data = Mrode11,
ginverse = list(calf = Ainv),
Gstart = matrix(0.2), Rstart = matrix(0.4),
maxit = 10, v = 2)
}
}
\references{
Mrode. 2005.
  Meyer & Smith. 1996.
  Gilmour et al. 1995.
  Thompson & Johnson. 1995.
}
\seealso{
\code{\link[MCMCglmm:MCMCglmm-package]{MCMCglmm}}
}
\author{
\strong{Maintainer}: Matthew Wolak \email{matthewwolak@gmail.com} (0000-0002-7962-0071)

}
