% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gremlin.R
\docType{package}
\name{gremlin-package}
\alias{gremlin-package}
\alias{_PACKAGE}
\title{Mixed-Effects REML Incorporating Generalized Inverses}
\description{
Fit linear mixed-effects models using restricted (or residual) maximum
likelihood (REML) and with generalized inverse matrices to specify covariance
structures for random effects. In particular, the package is suited to fit
quantitative genetic mixed models, often referred to as 'animal models'.
Implements the average information algorithm as the main tool to maximize the
restricted likelihood, but with other algorithms available.
}
\details{
The package also implements the average information algorithm to efficiently
maximize the log-likelihood (Thompson & Johnson 1995; Gilmour et al. 1995;
Meyer & Smith 1996). The average information algorithm combined with sparse
matrix techniques can potentially make model fitting very efficient.
}
\examples{
\dontrun{
  # Following the example from Mrode 2005, chapter 11.
  library(nadiv)  #<-- to construct inverse of the numerator relatedness matrix
  Ainv <- makeAinv(Mrode11[, 1:3])$Ainv

  gr11lmm <- gremlin(WWG11 ~ sex - 1,
random = ~ calf,
data = Mrode11,
ginverse = list(calf = Ainv),
Gstart = matrix(0.2), Rstart = matrix(0.4),  #<-- specify starting values
maxit = 15,    #<-- maximum iterations
     v = 2, vit = 1,  #<-- moderate screen output (`v`) every iteration (`vit`)
     algit = "AI")  #<-- only use Average Information algorithm iterations
  summary(gr11lmm)

  # Compare the model to a Linear Model with no random effects
  ## Use `update()` to update the model
  gr11lm <- update(gr11lmm, random = ~ 1)  #<-- `~1`=drop all random effects
  summary(gr11lm)

  # Do analysis of variance between the two models
  ## See AIC or evaluate likelihood ratio against a Chi-squared distribution
  anova(gr11lm, gr11lmm)
}
}
\references{
Mrode. 2005.
  Meyer & Smith. 1996.
  Gilmour et al. 1995.
  Thompson & Johnson. 1995.
}
\seealso{
\code{\link[MCMCglmm:MCMCglmm-package]{MCMCglmm}}
}
\author{
\strong{Maintainer}: Matthew Wolak \email{matthewwolak@gmail.com} (\href{https://orcid.org/0000-0002-7962-0071}{ORCID})

}
