% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Mrode11}
\alias{Mrode11}
\title{Weight gain data.}
\format{An object of class \code{data.frame} with 5 rows and 5 columns.}
\source{
Mrode. 2005.
}
\usage{
Mrode11
}
\description{
Data from chapter 11 in Mrode 2005. The variables are as follows:
}
\details{
\itemize{
  \item calf. a factor with levels \code{4} \code{5} \code{6} \code{7} \code{8}
  \item dam. a factor with levels \code{2} \code{5} \code{6}
  \item sire. a factor with levels \code{1} \code{3} \code{4}
  \item sex. a factor with levels \code{male} \code{female}
  \item WWG11. a numeric vector
}
}
\examples{
data(Mrode11)
}
\keyword{datasets}
