% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFuns.R
\name{stTrans}
\alias{stTrans}
\title{Transformation of starting parameters.}
\usage{
stTrans(x)
}
\arguments{
\item{x}{A \code{list} of starting parameters.}
}
\value{
A sparse \sQuote{dsCMatrix}
}
\description{
Transform start parameters into format \code{gremlin} expects.
}
\author{
\email{matthewwolak@gmail.com}
}
