% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformFuns.R
\name{vech2matlist}
\alias{vech2matlist}
\title{Vector to list of matrices.}
\usage{
vech2matlist(vech, skeleton)
}
\arguments{
\item{vech}{A \code{vector} of (co)variance parameters.}

\item{skeleton}{An example structure to map \code{vech} onto.}
}
\value{
A list of matrices of the same structure as \code{skeleton}.
}
\description{
Converts a vector of (co)variance parameters to a list of covariance matrices.
}
\author{
\email{matthewwolak@gmail.com}
}
